Move Extensions.System assembly into appropriate namespaces
