+++ Assets/Scripts/Animation/AnimationManager.cs
@@ -7,9 +7,9 @@ using PAC.Layers;
 using PAC.UI;
 using UnityEngine;
 using UnityEngine.Events;
-using PAC.Extensions;
 using PAC.Colour.Compositing;
 using PAC.Extensions.UnityEngine;
+using PAC.Extensions.System.Collections;
 
 namespace PAC.Animation
 {
