+++ Assets/Scripts/Data Structures/IntRange.cs
@@ -1,4 +1,4 @@
-using PAC.Extensions;
+using PAC.Extensions.System.Collections;
 using PAC.Interfaces;
 
 using System;
