+++ Assets/Scripts/Extensions/System/Collections/ICollectionExtensions.cs
@@ -1,14 +1,14 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace PAC.Extensions
+namespace PAC.Extensions.System.Collections
 {
     public static class ICollectionExtensions
     {
         /// <summary>
         /// Whether all the given elements are in the <see cref="ICollection{T}"/>.
         /// </summary>
-        public static bool ContainsAll<T>(this ICollection<T> iCollection, params T[] elements) => ContainsAll(iCollection, (IEnumerable<T>)elements);
+        public static bool ContainsAll<T>(this ICollection<T> iCollection, params T[] elements) => iCollection.ContainsAll((IEnumerable<T>)elements);
         /// <summary>
         /// Whether all the given elements are in the <see cref="ICollection{T}"/>.
         /// </summary>
@@ -17,7 +17,7 @@ namespace PAC.Extensions
         /// <summary>
         /// Whether at least one of the given elements is not in the <see cref="ICollection{T}"/>.
         /// </summary>
-        public static bool ContainsNotAll<T>(this ICollection<T> iCollection, params T[] elements) => ContainsNotAll(iCollection, (IEnumerable<T>)elements);
+        public static bool ContainsNotAll<T>(this ICollection<T> iCollection, params T[] elements) => iCollection.ContainsNotAll((IEnumerable<T>)elements);
         /// <summary>
         /// Whether at least one of the given elements is not in the <see cref="ICollection{T}"/>.
         /// </summary>
@@ -26,7 +26,7 @@ namespace PAC.Extensions
         /// <summary>
         /// Whether any of the given elements are in the <see cref="ICollection{T}"/>.
         /// </summary>
-        public static bool ContainsAny<T>(this ICollection<T> iCollection, params T[] elements) => ContainsAny(iCollection, (IEnumerable<T>)elements);
+        public static bool ContainsAny<T>(this ICollection<T> iCollection, params T[] elements) => iCollection.ContainsAny((IEnumerable<T>)elements);
         /// <summary>
         /// Whether any of the given elements are in the <see cref="ICollection{T}"/>.
         /// </summary>
@@ -35,7 +35,7 @@ namespace PAC.Extensions
         /// <summary>
         /// Whether none of the given elements are in the <see cref="ICollection{T}"/>.
         /// </summary>
-        public static bool ContainsNone<T>(this ICollection<T> iCollection, params T[] elements) => ContainsNone(iCollection, (IEnumerable<T>)elements);
+        public static bool ContainsNone<T>(this ICollection<T> iCollection, params T[] elements) => iCollection.ContainsNone((IEnumerable<T>)elements);
         /// <summary>
         /// Whether none of the given elements are in the <see cref="ICollection{T}"/>.
         /// </summary>
