+++ Assets/Scripts/Geometry/Shapes/Line.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 
 using PAC.DataStructures;
-using PAC.Extensions;
+using PAC.Extensions.System.Collections;
 using PAC.Geometry.Axes;
 using PAC.Geometry.Shapes.Interfaces;
