+++ Assets/Scripts/Geometry/Shapes/Path.cs
@@ -3,10 +3,10 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
-using PAC.Extensions;
 using PAC.Maths;
 using PAC.Geometry.Shapes.Interfaces;
 using PAC.Geometry.Axes;
+using PAC.Extensions.System.Collections;
 
 namespace PAC.Geometry.Shapes
 {
