+++ Assets/Scripts/Image Editing/ImageEdit.cs
@@ -1,7 +1,7 @@
 using System;
 
 using PAC.Exceptions;
-using PAC.Extensions;
+using PAC.Extensions.System.Collections;
 using PAC.Extensions.UnityEngine;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
