+++ Assets/Scripts/Image Editing/Outline.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-using PAC.Extensions;
+using PAC.Extensions.System.Collections;
 using PAC.Extensions.UnityEngine;
 using PAC.Geometry;
 using PAC.Geometry.Extensions;
