+++ Assets/Scripts/Layers/LayerManager.cs
@@ -3,7 +3,7 @@ using System.Linq;
 
 using PAC.Animation;
 using PAC.Colour.Compositing;
-using PAC.Extensions;
+using PAC.Extensions.System.Collections;
 using PAC.Files;
 using PAC.UI;
 using PAC.UndoRedo;
