+++ Assets/Scripts/Patterns/Extensions/IPattern2DExtensions.cs
@@ -1,6 +1,6 @@
 using System;
 
-using PAC.Extensions;
+using PAC.Extensions.System.Collections;
 using PAC.Extensions.UnityEngine;
 using PAC.Geometry;
