+++ Assets/Tests/Colour/Compositing/BlendMode_Tests.cs
@@ -4,7 +4,7 @@ using NUnit.Framework;
 
 using PAC.Colour;
 using PAC.Colour.Compositing;
-using PAC.Extensions;
+using PAC.Extensions.System.Collections;
 using PAC.Extensions.UnityEngine;
 using PAC.Json;
