+++ Assets/Tests/Drawing/CoordSnapping_Tests.cs
@@ -3,7 +3,7 @@ using System;
 using NUnit.Framework;
 
 using PAC.Drawing;
-using PAC.Extensions;
+using PAC.Extensions.System;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.Geometry.Shapes;
