+++ Assets/Tests/Extensions/System/IO/TextReaderExtensions_Tests.cs
@@ -1,7 +1,7 @@
 using NUnit.Framework;
-using PAC.Extensions;
 using System.IO;
 using System.Collections.Generic;
+using PAC.Extensions.System.IO;
 
 namespace PAC.Tests.Extensions
 {
