+++ Assets/Tests/Extensions/System/RandomExtensions_Tests.cs
@@ -4,7 +4,7 @@ using System.Linq;
 
 using NUnit.Framework;
 
-using PAC.Extensions;
+using PAC.Extensions.System;
 
 namespace PAC.Tests.Extensions
 {
