+++ Assets/Tests/Geometry/Shapes/IsometricCuboid_Tests.cs
@@ -1,6 +1,6 @@
 using NUnit.Framework;
 
-using PAC.Extensions;
+using PAC.Extensions.System;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.Geometry.Shapes;
